fix Saver
Q: If a creature with Saver is destroyed along at the same as a the specific race for the creatures saver ability, can you use the saver's ability?
A: Yes, you may. If this happens you may destroy a creature with saver and a creature of the specific race stays in the battle zone.
https://duelmasters.fandom.com/wiki/Saver#Rulings