fix for mana cost-related abilities
The current level of a card can't be used for an effect. For example, Frantic Chieftain shouldn't be able to return 1 of your creatures whose level has been reduced to 4 by Elf-X.

"While Cost Reduction and Cost Increase abilities alter the cost you pay, they don't actually increase or decrease the cost on the written card."
https://duelmasters.fandom.com/wiki/Mana_Cost#Rules

Added `Card.IsManaCost` to check if a card's mana cost is equal to a given value.
Added `Card.IsManaCostBelow` to check if a card's mana cost is less than or equal to a given value.
Added `Card.IsManaCostAbove` to check if a card's mana cost is greater than or equal to a given value.

Renamed `Card.GetManaCost` to `Card.GetPlayCost`.

`Card.GetManaCost` now returns the original level of a card.