replace ability functions
Replaced the following trigger ability functions with `dm.AddSingleTriggerEffectCustom`:
- `dm.AddSingleBlockTriggerEffect`

Replaced the following trigger ability functions with `dm.AddTriggerEffectCustom`:
- `dm.AddTurnEndTriggerEffect`
- `dm.AddTurnStartTriggerEffect`
- `dm.AddEnterGraveTriggerEffect`

Replaced the following static ability functions with `dm.AddSingleGrantEffectCustom`:
- `dm.AddStaticEffectSingleAttackTrigger`
- `dm.AddStaticEffectSingleComeIntoPlay`
- `dm.AddBlockTriggerEffect`

Added `dm.PreviousLocationMZoneCheck` to check if a card was put into the graveyard (banished face-up) from the mana zone (graveyard or as a face-down banished card). This function is required due to the workaround of flipping a face-down banished card face-up.